+++ src/php/Run/RunFacadeInterface.php
@@ -46,4 +46,9 @@ interface RunFacadeInterface
     public function writeLocatedException(OutputInterface $output, CompilerException $e): void;
 
     public function writeStackTrace(OutputInterface $output, Throwable $e): void;
+
+    /**
+     * @return list<NamespaceInformation>
+     */
+    public function getLoadedNamespaces(): array;
 }