+++ tests/php/Unit/Printer/TypePrinter/ArrayPrinterTest.php
@@ -29,28 +29,28 @@ final class ArrayPrinterTest extends TestCase
     public static function providerPrint(): Generator
     {
         yield 'Empty array' => [
-            'form' => [],
-            'expected' => '<PHP-Array []>',
+            [],
+            '<PHP-Array []>',
         ];
 
         yield 'simple numeric list' => [
-            'form' => [1, 2, 3],
-            'expected' => '<PHP-Array [1, 2, 3]>',
+            [1, 2, 3],
+            '<PHP-Array [1, 2, 3]>',
         ];
 
         yield 'simple string list' => [
-            'form' => ['s1', 's2', 's3'],
-            'expected' => '<PHP-Array [s1, s2, s3]>',
+            ['s1', 's2', 's3'],
+            '<PHP-Array [s1, s2, s3]>',
         ];
 
         yield 'key-value dictionary with numbers' => [
-            'form' => ['k1' => 1, 'k2' => 2],
-            'expected' => '<PHP-Array [k1:1, k2:2]>',
+            ['k1' => 1, 'k2' => 2],
+            '<PHP-Array [k1:1, k2:2]>',
         ];
 
         yield 'key-value dictionary with strings' => [
-            'form' => ['k1' => 'v1', 'k2' => 'v2'],
-            'expected' => '<PHP-Array [k1:v1, k2:v2]>',
+            ['k1' => 'v1', 'k2' => 'v2'],
+            '<PHP-Array [k1:v1, k2:v2]>',
         ];
     }
 }