+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 
 ## Unreleased
 
+- Add functions
+  - `str/pad-left`
+  - `str/pad-right`
+  - `trace/dotrace`
+  - `trace/dbg`
+- Add `phel ns [inspect]` command
+
+## [0.17.0](https://github.com/phel-lang/phel-lang/compare/v0.16.1...v0.17.0) - 2025-06-01
+
 - Fix `php/echo` does not work (#729)
 - Fix `hash-map` malfunction notice (#786)
 - Fix header injection vulnerability (#803)
@@ -31,6 +40,7 @@ All notable changes to this project will be documented in this file.
 - Add `phel->php` and `php->phel` functions (#829)
 - Add `ApiFacade::replComplete` service (#830)
 - Add `str\contains?` function (#833)
+- Add display deprecated metadata on doc (#834)
 - Improve vector performance (#823)
 - Improve `DependenciesForNamespace` performance
 - Use PHPStan level 2